+++ postgres/mod.ts
@@ -1,3 +1,4 @@
 export * from "./postgres-migrator.ts";
 export * from "./postgres-client.ts";
+export * from "./postgres-js.ts";
 export * from "./tables.ts";